Authorization with Gate

+++ app/Http/Controllers/ArtistController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Artist;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ArtistController extends Controller
 {
@@ -25,6 +26,9 @@ class ArtistController extends Controller
      */
     public function create()
     {
+        if (! Gate::allows('create-artist')) {
+            abort(403);
+        }
         return view('artist.create');
 
     }
@@ -70,6 +74,9 @@ class ArtistController extends Controller
      */
     public function edit(string $id)
     {
+        if (! Gate::allows('create-artist')) {
+            abort(403);
+        }
         $artist = Artist::find($id);
 
         return view('artist.edit',[
@@ -104,6 +111,9 @@ class ArtistController extends Controller
      */
     public function destroy(string $id)
     {
+        if (! Gate::allows('delete-artist')) {
+            abort(403);
+        }
         Artist::destroy($id);
 
         return redirect()->route('artist.index');
